Debug

+++ app/matching_algorithm.py
@@ -153,13 +153,13 @@ def propose_restaurants(candidates, group_preferences, meeting_id):
                 continue
             else:
                 budget_score = 1 - (restaurant.end_price / group_preferences['max_budget_per_person'][0])
-                score += float(group_preferences['max_budget_per_person'][1] * budget_score)
+                score += float(group_preferences['max_budget_per_person'][1]) * float(budget_score)
 
             # 2. Rating (normalized 0-1 scale)
             if restaurant.rating < group_preferences['rating'][0]:
                 continue
             else:
-                score += float(group_preferences['rating'][1] * (restaurant.rating / 5))
+                score += float(group_preferences['rating'][1]) * float(restaurant.rating / 5)
 
             # 3. Distance from centroid (closer = better)
             #if 'distance_from_centroid' in restaurant:
